feat(rust): OAuth-Callback komplett + Halbport-Audit — 4 stille Lücken im Raid-Tracking geschlossen (#145)
OAuth-Callback (Code-komplett, bewusst noch proxied):
- tb-transport-twitch: exchange_user_code (authorization_code-Grant mit
  redirect_uri) + fetch_token_owner (GET /helix/users mit User-Bearer —
  der fehlende Identitäts-Schritt) inkl. Wiremock-Tests
- TwitchTokenClient-Trait + token_owner(); HelixTokenClient-Stub
  ('läuft über Python-Dashboard') durch echte Impl ersetzt,
  redirect_uri als Konstruktor-Feld
- raid_oauth_impl::oauth_callback voller Python-Port: Mismatch-Checks
  (expected_user_id/login, discord:/public:-Ausnahmen), Scope-Profil-
  Check, has_saved_auth_record, AuthWriter-Persist, redirect_url,
  generische 500-Failure-Texte der internen API; 7 Integrationstests
  (Erfolg/Mismatches/Scopes/Exchange-Fail/Bad-State, je mit
  Nichts-gespeichert-Prüfung)
- Handler auf geteilten Idempotenz-Layer (Ergebnis cacht als HTTP 200
  wie Python, raid.py:292); OAuthCallbackResult + redirect_url
- Flip-Bedingung dokumentiert: complete_setup/sync_partner_state-
  Followups (Moderator-Setup, Chat-Begrüßung, EventSub-Sofort-Sub)

Halbport-Audit (84 Agenten, 3-fach verifiziert) — Live-Fixes:
- raid_arrival_wiring: record_secondary_signal war No-Op →
  find_recent_arrival (600s-Fenster, Python-TTL) + Signal-Merge via
  tb_raid::serialize_confirmation_signals + update_arrival
- store_orphan_chat_notification war No-Op → In-Memory-Orphan-Store
  (15s Grace / 900s Retention wie Python) + Sweeper-Task (5s) +
  Promotion als independent arrival (classify + record_arrival +
  Suppression-Mark 180s) + Pop bei Confirm-Korrelation
- track_confirmed-Result wurde mit let _ verworfen → error-Log
- main.rs: periodischer Voll-Refresh aller Partner-Scores (300s,
  Python maybe_schedule_partner_raid_score_reconciliation) — erster
  Tick live verifiziert (52 Scores frisch)

Typ-Drift-Nachzieher (Prod-Schema-Dump als Referenz):
- RecentSessionRow: duration_seconds/peak_viewers/follower_delta
  ::BIGINT-Casts (INT4 in Prod); OverviewMetricsRow MAX()::BIGINT
- archive_streamer: admin_archived_at als String (TEXT in Prod, nicht
  TIMESTAMPTZ); Test-DDLs an Prod angeglichen (twitch_partners
  manual_verified_* TEXT, sessions INT4, reauth_notified_at TIMESTAMPTZ,
  raids.rs Schema-Drop statt IF NOT EXISTS)

tb-dashboard-api (latent, Welle D):
- recent_raids: WHERE success = TRUE + LIMIT 10 (api_public.py-Parität)
- CorsLayer::permissive nur noch auf Public-Routen (wie Python)
- stats_handler-Docblock entirreführt; eventsub-KPI-Platzhalter markiert
- 04-cutover-plan: Stand 11.6. abends + Welle-D-Blocker-Inventar

+++ rust/crates/tb-analytics/src/overview.rs
@@ -26,7 +26,7 @@ pub async fn overview_metrics(
         r#"
         SELECT
             AVG(s.avg_viewers)::FLOAT8                                AS avg_avg_viewers,
-            MAX(s.peak_viewers)                                       AS max_peak_viewers,
+            MAX(s.peak_viewers)::BIGINT                               AS max_peak_viewers,
             SUM(s.avg_viewers * s.duration_seconds / 3600.0)::FLOAT8  AS total_hours_watched,
             SUM(s.duration_seconds / 3600.0)::FLOAT8                  AS total_airtime_hours,
             SUM(CASE
